… source/jormungandr/jormungandr/scenarios/tests/journey_compare_tests.py
… source/jormungandr/jormungandr/scenarios/utils.py
… source/jormungandr/jormungandr/utils.py
… source/jormungandr/tests/journey_common_tests.py