https://jira.kisio.org/browse/NAVP-861

I used the following request, with which kraken is called  99 times (i requested 100 journeys) for the benchmark
```bash
time curl 'http://localhost:5000/v1/coverage/stif/journeys?from=stop_area%3A0%3ASA%3A8768600&to=stop_area%3A0%3ASA%3A8775860&datetime=20180423T083802&count=100' > /dev/null
```
In this [loop:](https://github.com/CanalTP/navitia/pull/2404/files#diff-f050c03e2b24acf6b927fade7d249a87L813):


|   time used in   | C++(Kraken) | Python |
|--------|--------|--------|
| Before |    4.46s    | 3.44s        | 
| After  |      4.41s  |  0.13s      | 

another request, with which kraken is called for 30 times (100 journeys are requested, same as before)
```bash
time curl 'http://localhost:5000/v1/coverage/stif/journeys?from=2.284794%3B48.911762&to=2.37715%3B48.846781&count=100&datetime=20180423T083802&' > /dev/null
```

|   time used in     | C++(Kraken) | Python |
|--------|--------|--------|
| Before |   3.21s   | 1.2s        | 
| After  |     3.30s  |  0.15s      |